+++ src/backend/joanie/lms_handler/backends/dummy.py
@@ -75,3 +75,35 @@ class DummyLMSBackend(BaseLMSBackend):
             cache.set(cache_key, True)
         else:
             cache.delete(cache_key)
+
+    def get_grades(self, username, resource_link):
+        """
+        Get a fake user's grade for a course run given its resource_link.
+
+        The return dict looks like a grade summary of a course run which has only one
+        graded exercice called "Final Exam" which have a grade of 0.0.
+        """
+        return {
+            "passed": False,
+            "grade": None,
+            "percent": 0.0,
+            "totaled_scores": {
+                "Final Exam": [[0.0, 1.0, True, "First section", None]],
+            },
+            "grade_breakdown": [
+                {
+                    "category": "Final Exam",
+                    "percent": 0.0,
+                    "detail": "Final Exam = 0.00% of a possible 0.00%",
+                }
+            ],
+            "section_breakdown": [
+                {
+                    "category": "Final Exam",
+                    "prominent": True,
+                    "percent": 0.0,
+                    "detail": "Final Exam = 0%",
+                    "label": "FE",
+                },
+            ],
+        }